api-server/requirements.txt
click==6.7
Flask==0.12.2
Flask-Cors==3.0.3
Flask-MySQLdb==0.2.0
Flask-SQLAlchemy==2.3.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
PyJWT==1.6.0
six==1.11.0
SQLAlchemy==1.2.5
Werkzeug==0.14.1
